댓글 삭제하기 진행중

+++ lib/util/dio_logging_interceptor.dart
@@ -58,41 +58,43 @@ class DioLoggingInterceptors extends dio.Interceptor {
     print(
         "${dioError.response != null ? dioError.response!.data : 'Unknown Error'}");
     print("<-- End error");
+    if (dioError.response != null) {
+      int responseCode = dioError.response!.statusCode!;
+      String? oldAccessToken = await _secureStorage.read(key: 'access_token');
+      if (oldAccessToken != null &&
+          (responseCode == 401 || responseCode == 422)) {
+        _dio.interceptors.requestLock.lock();
+        _dio.interceptors.responseLock.lock();
+        String? userId = await _secureStorage.read(key: 'user_id');
+        User? currentUser = _auth.currentUser;
+        dio.Response? resp;
+        if (userId == null && currentUser != null) {
+          resp = await Dio(dio.BaseOptions(
+            baseUrl: 'http://34.64.204.217/api/',
+            connectTimeout: 5000,
+            receiveTimeout: 3000,
+          )).post('auth/authenticated/access-token',
+              data: {'email': currentUser.email});
+        } else if (userId != null && currentUser == null) {
+          resp = await Dio(BaseOptions(
+            baseUrl: 'http://34.64.204.217/api/',
+            connectTimeout: 5000,
+            receiveTimeout: 3000,
+          )).post('auth/unauthenticated/access-token',
+              data: {'user_id': userId});
+        }
+        //ToDO: 고쳐야됨
+        final token = Token.fromJson(resp!.data['access_token']);
+        String newAccessToken = token.accessToken;
+        await _secureStorage.delete(key: 'access_token');
+        await _secureStorage.write(key: 'access_token', value: newAccessToken);
 
-    int responseCode = dioError.response!.statusCode!;
-    String? oldAccessToken = await _secureStorage.read(key: 'access_token');
-    if (oldAccessToken != null &&
-        (responseCode == 401 || responseCode == 422)) {
-      _dio.interceptors.requestLock.lock();
-      _dio.interceptors.responseLock.lock();
-      String? userId = await _secureStorage.read(key: 'user_id');
-      User? currentUser = _auth.currentUser;
-      dio.Response? resp;
-      if (userId == null && currentUser != null) {
-        resp = await Dio(dio.BaseOptions(
-          baseUrl: 'http://34.64.204.217/api/',
-          connectTimeout: 5000,
-          receiveTimeout: 3000,
-        )).post('auth/authenticated/access-token',
-            data: {'email': currentUser.email});
-      } else if (userId != null && currentUser == null) {
-        resp = await Dio(BaseOptions(
-          baseUrl: 'http://34.64.204.217/api/',
-          connectTimeout: 5000,
-          receiveTimeout: 3000,
-        )).post('auth/unauthenticated/access-token', data: {'user_id': userId});
+        dio.RequestOptions requestOptions = dioError.response!.requestOptions;
+        requestOptions.headers.addAll({'requirestoken': true});
+        _dio.interceptors.requestLock.unlock();
+        _dio.interceptors.responseLock.unlock();
+        return _retry(requestOptions);
       }
-      //ToDO: 고쳐야됨
-      final token = Token.fromJson(resp!.data['access_token']);
-      String newAccessToken = token.accessToken;
-      await _secureStorage.delete(key: 'access_token');
-      await _secureStorage.write(key: 'access_token', value: newAccessToken);
-
-      dio.RequestOptions requestOptions = dioError.response!.requestOptions;
-      requestOptions.headers.addAll({'requirestoken': true});
-      _dio.interceptors.requestLock.unlock();
-      _dio.interceptors.responseLock.unlock();
-      return _retry(requestOptions);
     } else {
       super.onError(dioError, handler);
     }
